filter with productName and localization works, filter categoryName works without price range

+++ app/core/src/main/java/com/teamg/tourdeshot/core/api/local/filter/ProductCategory.java
@@ -9,8 +9,8 @@ public class ProductCategory {
 
     private String id;
 
-    private BigDecimal priceFrom;
+    private Double priceFrom;
 
-    private BigDecimal priceTo;
+    private Double priceTo;
 
 }
